lnop/bumblebee: add users from some groups to bumblebee by default
Discussed with ArchangeGabriel (Bruno Pagani), it seems to be more
Ubuntu-friendly to require less steps from the user. On a default
installation, the groups sudo, adm and admin usually contains the desktop users
who are likely going to use bumblebee. To account for environments where the
administrator wants more control, users are not added to already created groups
and to a different group than configured in /etc/bumblebee/bumblebee.conf. In
addition, no group is created if the group in /etc/bumblebee/bumblebee.conf is
different than the default.